refactor: userTypeSelectView extract
